requirements.txt
Flask>=3.0.0,<4.0.0
Werkzeug>=3.0.0,<4.0.0
Jinja2>=3.1.0,<4.0.0
numpy>=1.24.0,<3.0.0
pandas>=2.0.0,<3.0.0
scipy>=1.10.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0
joblib>=1.3.0,<2.0.0
xgboost>=2.0.0,<4.0.0
lightgbm>=4.0.0,<5.0.0
tensorflow>=2.15.0,<3.0.0
keras>=3.0.0,<4.0.0
h5py>=3.10.0,<4.0.0
Pillow>=10.0.0,<13.0.0
python-multipart>=0.0.6,<1.0.0
requests>=2.31.0,<3.0.0
